+++ src/settings/ActionProfiles/ActionProfiles.js
@@ -141,6 +141,8 @@ export class ActionProfiles extends Component {
     initialValues: {
       name: '',
       description: '',
+      action: 'CREATE',
+      folioRecord: 'ITEM',
     },
     RecordView: ViewActionProfile,
     RecordForm: ActionProfilesForm,