+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add tags accordion for match profiles (UIDATIMP-16)
 * Add tags accordion for action profiles (UIDATIMP-17)
 * Create data import settings page's 3rd pane for Field Mapping Profiles (UIDATIMP-9)
+* Implement details view for mapping profiles (UIDATIMP-219)
 
 ## [1.3.0](https://github.com/folio-org/ui-data-import/tree/v1.3.0) (2019-06-12)
 * Create data import settings page's 3rd pane for Match Profiles. Add string capitalization (capitalize) and HTML special chars decoding (htmlDecode) utils (UIDATIMP-6)